Fix all clippy warnings
- seq.rs: allow `should_implement_trait` on `SeqNo::add(n: u32)` — the
  standard `Add` trait has a different signature so implementing it would
  change the call sites.
- codec.rs: allow `too_many_arguments` on `encode_data` (structural, not
  worth wrapping yet); simplify localhost peer_ip literal from the
  no-op `127|(0<<8)|(0<<16)|(1<<24)` to `127`.
- send_buffer.rs: allow `result_unit_err` on `add` — the unit error is
  intentional (buffer-full signal, no detail needed).
- connection.rs: allow `too_many_arguments` on `new_connected` (listener
  factory, called from one place).
- listener.rs: box `Connection` inside `ListenerOutput::Accept` to
  eliminate the large-enum-variant warning (Connection is ~1.4 KB).
  Update the matching arm in tokio_impl.rs to dereference the box.
- tokio_impl.rs: use `io::Error::other()` instead of the verbose
  `io::Error::new(io::ErrorKind::Other, …)` form; collapse the
  nested listener-stale check into a single `if let … && …`.
- ledbat.rs: remove unnecessary parentheses around `ctx.rtt_us / 2`.
- udt-async/src/lib.rs: gate `MAX_PAYLOAD_SIZE` behind
  `#[cfg(feature = "tokio")]` so the crate compiles without features.

+++ udt-proto/src/congestion/ledbat.rs
@@ -117,7 +117,7 @@ impl CongestionControl for Ledbat {
     }
 
     fn on_ack(&mut self, _ack_seq: SeqNo, ctx: CcContext) -> CcOutput {
-        let delay_us = (ctx.rtt_us / 2) as u32;
+        let delay_us = ctx.rtt_us / 2;
         self.update_base_delay(delay_us, ctx.now_us);
         self.update_current_delay(delay_us);
